+++ src/47-permuteUnique.cpp
@@ -21,7 +21,7 @@ private:
             res.emplace_back(path);
             return;
         }
-        for (int i = 0; i < nums.size(); ++i) {
+        for (int i = 0; i < static_cast<int>(nums.size()); ++i) {
             if (used[i]) continue;
             if (i > 0 && nums[i] == nums[i - 1] && !used[i - 1]) continue;
             used[i] = true;
