+++ src/5-longestPalindrome.cpp
@@ -10,7 +10,7 @@ class Solution {
 public:
     string longestPalindrome(string s) {
         pair<int, int> index{0, 0};
-        for (int i = 0; i < s.size(); i++) {
+        for (int i = 0; i < static_cast<int>(s.size()); i++) {
             auto index1 = findPalindrome(s, i, i);
             auto index2 = findPalindrome(s, i, i + 1);
             index = index1.second - index1.first > index.second - index.first ? index1 : index;
@@ -21,7 +21,7 @@ public:
 
 private:
     static pair<int, int> findPalindrome(string s, int left, int right) {
-        while (left >= 0 && right < s.size() && s[left] == s[right]) {
+        while (left >= 0 && right < static_cast<int>(s.size()) && s[left] == s[right]) {
             left--;
             right++;
         }
