+++ src/53-maxSubArray.cpp
@@ -11,7 +11,7 @@ public:
     int maxSubArray(vector<int>& nums) {
         int res = INT_MIN;
         int sum = 0;
-        for (int i = 0; i < nums.size(); i++) {
+        for (int i = 0; i < static_cast<int>(nums.size()); i++) {
             sum += nums[i];
             res = max(res, sum);
             if (sum < 0) {
