+++ src/66-plusOne.cpp
@@ -9,7 +9,7 @@ using namespace std;
 class Solution {
 public:
     vector<int> plusOne(vector<int>& digits) {
-        for (int i = digits.size() - 1; i >= 0; --i) {
+        for (int i = static_cast<int>(digits.size()) - 1; i >= 0; --i) {
             if (digits[i] < 9) {
                 digits[i]++;
                 return digits;
